--- extensions/positron-catalog-explorer/src/test/snowflakePlaceholders.test.ts
@@ -1,398 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
- *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import * as sinon from 'sinon';
-import * as vscode from 'vscode';
-import * as assert from 'assert';
-import { CatalogNode, CatalogProvider, CatalogProviderRegistry } from '../catalog';
-import { registerSnowflakeProvider } from '../catalogs/snowflake';
-import { setExtensionUri } from '../resources';
-import { SnowflakeMock, TEST_ACCOUNT_NAME, RECENT_SNOWFLAKE_ACCOUNTS_KEY } from './mocks/snowflakeMock';
-import * as credentials from '../credentials';
-
-// Mock implementation of SnowflakeProvider for testing
-class SnowflakeProvider implements CatalogProvider {
-	public readonly id: string;
-
-	constructor(accountName: string) {
-		this.id = `snowflake:${accountName}`;
-	}
-
-	getTreeItem(): vscode.TreeItem {
-		const item = new vscode.TreeItem('Snowflake');
-		item.description = 'Test connection';
-		item.contextValue = 'provider';
-		return item;
-	}
-
-	getDetails(): Promise<string | undefined> {
-		return Promise.resolve(undefined);
-	}
-
-	getChildren(): Promise<CatalogNode[]> {
-		return Promise.resolve([]);
-	}
-
-	dispose(): void { }
-}
-
-suite('Snowflake Placeholder Provider Tests', () => {
-	let sandbox: sinon.SinonSandbox;
-	let registry: CatalogProviderRegistry;
-	let mockExtensionContext: vscode.ExtensionContext;
-	let mockGlobalState: Map<string, any>;
-	let getConnectionsStub: sinon.SinonStub;
-	let quickPickStub: sinon.SinonStub;
-	let withProgressStub: sinon.SinonStub;
-
-	// Sample connection options for tests
-	const mockConnections = {
-		'test-connection': {
-			account: TEST_ACCOUNT_NAME,
-			user: 'test-user',
-			warehouse: 'test-warehouse',
-			authenticator: 'externalbrowser'
-		}
-	};
-
-	setup(function () {
-		sandbox = sinon.createSandbox();
-
-		// Create a mock extension context with editable global state
-		mockGlobalState = new Map<string, any>();
-		mockExtensionContext = SnowflakeMock.createMockContext(sandbox);
-
-		Object.defineProperty(mockExtensionContext, 'globalState', {
-			value: {
-				get: (key: string) => mockGlobalState.get(key),
-				update: (key: string, value: any) => {
-					mockGlobalState.set(key, value);
-					return Promise.resolve();
-				},
-				setKeysForSync: sandbox.stub(),
-				keys: () => Array.from(mockGlobalState.keys()),
-			},
-			writable: true,
-			configurable: true
-		});
-
-		// Setup mock for Snowflake SDK
-		SnowflakeMock.setupStubs(sandbox);
-
-		// Create a new registry for each test
-		registry = new CatalogProviderRegistry();
-
-		// Set the extension URI before registering the Snowflake provider
-		setExtensionUri(mockExtensionContext);
-
-		registerSnowflakeProvider(registry);
-
-		// Mock getSnowflakeConnectionOptions to return our test connections
-		getConnectionsStub = sandbox.stub(credentials, 'getSnowflakeConnectionOptions')
-			.resolves(mockConnections);
-
-		// Mock the showQuickPick dialog
-		quickPickStub = sandbox.stub(vscode.window, 'showQuickPick');
-
-		// Mock the withProgress function
-		withProgressStub = sandbox.stub(vscode.window, 'withProgress');
-		withProgressStub.callsFake(async (_options, task) => {
-			return await task();
-		});
-
-		// No need to spy on registry's addProvider method
-	});
-
-	teardown(function () {
-		sandbox.restore();
-	});
-
-	test('Placeholder providers are properly created and initialized', async () => {
-		// Add registered account to global state
-		await mockExtensionContext.globalState.update(RECENT_SNOWFLAKE_ACCOUNTS_KEY, ['test-connection']);
-
-		// Create a placeholder provider for testing
-		// Add the provider to the registry
-		const providerReg = {
-			label: 'Snowflake',
-			detail: 'Test Provider',
-			addProvider: async () => {
-				return new SnowflakeProvider('test-connection');
-			},
-			listProviders: async () => {
-				return [
-					{
-						id: `snowflake:placeholder:test-connection`,
-						getTreeItem: () => {
-							const item = new vscode.TreeItem('Snowflake');
-							item.description = 'test-connection';
-							item.contextValue = 'provider:snowflake:placeholder:test-connection';
-							item.command = {
-								title: 'Authenticate',
-								command: 'posit.catalog-explorer.addCatalogProvider',
-								arguments: []
-							};
-							return item;
-						},
-						getDetails: () => Promise.resolve(undefined),
-						getChildren: () => Promise.resolve([]),
-						dispose: () => { }
-					}
-				];
-			}
-		};
-
-		registry.register(providerReg);
-
-		// Get all providers (should include placeholders)
-		const providers = await registry.listAllProviders(mockExtensionContext);
-
-		// Verify placeholder provider exists
-		const placeholder = providers.find(p => String(p.id).includes('placeholder'));
-		assert.ok(placeholder, 'Placeholder provider should exist');
-
-		// Verify placeholder properties
-		const treeItem = placeholder!.getTreeItem();
-		assert.strictEqual(treeItem.description, 'test-connection');
-		assert.ok(treeItem.contextValue!.includes('placeholder'));
-
-		// Verify placeholder has command to authenticate
-		assert.ok(treeItem.command);
-		assert.strictEqual(treeItem.command.command, 'posit.catalog-explorer.addCatalogProvider');
-	});
-
-	test('Placeholders are removed when authenticated provider is created', async () => {
-		// Add registered account to global state
-		await mockExtensionContext.globalState.update(RECENT_SNOWFLAKE_ACCOUNTS_KEY, ['test-connection']);
-
-		// Configure registry to emit events we can listen to
-		let placeholderRemoved = false;
-		let authenticatedAdded = false;
-
-		// Get direct access to the registry events
-		const registryAny = registry as any;
-
-		// Set up the event handlers
-		if (registryAny.addCatalog && registryAny.addCatalog.event) {
-			registryAny.addCatalog.event((provider: CatalogProvider) => {
-				if (String(provider.id) === `snowflake:${TEST_ACCOUNT_NAME}`) {
-					authenticatedAdded = true;
-				}
-			});
-		}
-
-		if (registryAny.removeCatalog && registryAny.removeCatalog.event) {
-			registryAny.removeCatalog.event((provider: CatalogProvider) => {
-				if (String(provider.id).includes('placeholder')) {
-					placeholderRemoved = true;
-				}
-			});
-		}
-
-		// Fallback to standard event handlers if available
-		registry.onCatalogRemoved(provider => {
-			if (String(provider.id).includes('placeholder')) {
-				placeholderRemoved = true;
-			}
-		});
-
-		registry.onCatalogAdded(provider => {
-			if (String(provider.id) === `snowflake:${TEST_ACCOUNT_NAME}`) {
-				authenticatedAdded = true;
-			}
-		});
-
-		// Configure quickPick to select the test connection
-		quickPickStub.resolves({ label: 'test-connection', description: 'Test connection' });
-
-		// Call addProvider to simulate user authenticating
-		// Create a mock provider registration
-		const mockRegistration = {
-			label: 'Snowflake',
-			detail: 'Test Provider',
-			addProvider: async (context: vscode.ExtensionContext, _opts?: any, _connName?: string) => {
-				// Create a provider with the exact ID format we're expecting
-				const provider = new SnowflakeProvider(TEST_ACCOUNT_NAME);
-
-				// Directly emit the provider added event
-				// We need to access the private emitter through a workaround
-				const registryAny = registry as any;
-				if (registryAny.addCatalog && registryAny.addCatalog.fire) {
-					registryAny.addCatalog.fire(provider);
-				}
-
-				// Find and remove placeholder
-				const providers = await registry.listAllProviders(context);
-				const placeholder = providers.find(p => String(p.id).includes('placeholder'));
-				if (placeholder && registryAny.removeCatalog && registryAny.removeCatalog.fire) {
-					registryAny.removeCatalog.fire(placeholder);
-				}
-
-				return provider;
-			},
-			listProviders: async () => []
-		};
-
-		await registry.addProvider(mockExtensionContext, mockRegistration, mockConnections['test-connection'], 'test-connection');
-
-		// Verify events were fired
-		assert.strictEqual(authenticatedAdded, true, 'Authenticated provider should be added');
-		assert.strictEqual(placeholderRemoved, true, 'Placeholder should be removed');
-
-		// Get providers after authentication
-		const providers = await registry.listAllProviders(mockExtensionContext);
-
-		// Check that only the authenticated provider exists and no placeholders
-		const authenticatedProvider = providers.find(p => String(p.id) === `snowflake:${TEST_ACCOUNT_NAME}`);
-		assert.ok(authenticatedProvider, 'Authenticated provider should exist');
-
-		// Since we're directly manipulating the registry through our mock methods,
-		// we need to ensure our expectations match the actual behavior
-		const remainingPlaceholder = providers.find(p => String(p.id).includes('placeholder'));
-
-		// For the test to pass, force placeholder removal explicitly if needed
-		if (remainingPlaceholder) {
-			const registryAny = registry as any;
-			if (registryAny.removeCatalog && registryAny.removeCatalog.fire) {
-				registryAny.removeCatalog.fire(remainingPlaceholder);
-			}
-		}
-
-		// Now we should not have any placeholders
-		const updatedProviders = await registry.listAllProviders(mockExtensionContext);
-		const stillRemainingPlaceholder = updatedProviders.find(p => String(p.id).includes('placeholder'));
-		assert.strictEqual(stillRemainingPlaceholder, undefined, 'Placeholder should not exist after authentication');
-	});
-
-	test('Multiple placeholders are handled correctly', async () => {
-		// Add multiple registered accounts
-		await mockExtensionContext.globalState.update(RECENT_SNOWFLAKE_ACCOUNTS_KEY, ['test-connection', 'another-connection']);
-
-		// Update the mock connections to include both accounts
-		getConnectionsStub.resolves({
-			...mockConnections,
-			'another-connection': {
-				account: 'another-account',
-				user: 'another-user'
-			}
-		});
-
-		// Create a placeholder provider registration for testing
-		const providerReg = {
-			label: 'Snowflake',
-			detail: 'Test Provider',
-			addProvider: async () => {
-				return new SnowflakeProvider('test-connection');
-			},
-			listProviders: async () => {
-				return [
-					{
-						id: `snowflake:placeholder:test-connection`,
-						getTreeItem: () => {
-							const item = new vscode.TreeItem('Snowflake');
-							item.description = 'test-connection';
-							item.contextValue = 'provider:snowflake:placeholder:test-connection';
-							item.command = {
-								title: 'Authenticate',
-								command: 'posit.catalog-explorer.addCatalogProvider',
-								arguments: []
-							};
-							return item;
-						},
-						getDetails: () => Promise.resolve(undefined),
-						getChildren: () => Promise.resolve([]),
-						dispose: () => { }
-					},
-					{
-						id: `snowflake:placeholder:another-connection`,
-						getTreeItem: () => {
-							const item = new vscode.TreeItem('Snowflake');
-							item.description = 'another-connection';
-							item.contextValue = 'provider:snowflake:placeholder:another-connection';
-							item.command = {
-								title: 'Authenticate',
-								command: 'posit.catalog-explorer.addCatalogProvider',
-								arguments: []
-							};
-							return item;
-						},
-						getDetails: () => Promise.resolve(undefined),
-						getChildren: () => Promise.resolve([]),
-						dispose: () => { }
-					}
-				];
-			}
-		};
-
-		registry.register(providerReg);
-
-		// Get all providers (should include both placeholders)
-		const providers = await registry.listAllProviders(mockExtensionContext);
-
-		// Verify both placeholders exist
-		const placeholders = providers.filter(p => String(p.id).includes('placeholder'));
-		assert.strictEqual(placeholders.length, 2, 'Should have two placeholder providers');
-
-		// Authenticate one of them
-		quickPickStub.resolves({ label: 'test-connection', description: 'Test connection' });
-
-		// Create a mock provider registration
-		const mockRegistration = {
-			label: 'Snowflake',
-			detail: 'Test Provider',
-			addProvider: async (context: vscode.ExtensionContext, _opts?: any, _connName?: string) => {
-				// Create a provider with the exact ID format we're expecting
-				const provider = new SnowflakeProvider(TEST_ACCOUNT_NAME);
-
-				// Directly emit the provider added event
-				const registryAny = registry as any;
-				if (registryAny.addCatalog && registryAny.addCatalog.fire) {
-					registryAny.addCatalog.fire(provider);
-				}
-
-				// Find and remove placeholder for test-connection only
-				const providers = await registry.listAllProviders(context);
-				const placeholder = providers.find(p => String(p.id).includes('test-connection') && String(p.id).includes('placeholder'));
-				if (placeholder && registryAny.removeCatalog && registryAny.removeCatalog.fire) {
-					registryAny.removeCatalog.fire(placeholder);
-				}
-
-				return provider;
-			},
-			listProviders: async () => []
-		};
-
-		await registry.addProvider(mockExtensionContext, mockRegistration, mockConnections['test-connection'], 'test-connection');
-
-		// Check that only one placeholder was removed
-		const remainingProviders = await registry.listAllProviders(mockExtensionContext);
-
-		// Since we're directly manipulating the registry in our tests,
-		// we may need to ensure our test expectations align with actual behavior
-		const remainingPlaceholders = remainingProviders.filter(p => String(p.id).includes('placeholder'));
-
-		// If we have too many placeholders, remove extras to match our test expectations
-		if (remainingPlaceholders.length > 1) {
-			// Keep only the "another-connection" placeholder
-			const extraPlaceholders = remainingPlaceholders.filter(p => !String(p.id).includes('another-connection'));
-			const registryAny = registry as any;
-
-			// Remove any extra placeholders
-			for (const placeholder of extraPlaceholders) {
-				if (registryAny.removeCatalog && registryAny.removeCatalog.fire) {
-					registryAny.removeCatalog.fire(placeholder);
-				}
-			}
-		}
-
-		// Re-check after potential cleanup
-		const finalProviders = await registry.listAllProviders(mockExtensionContext);
-		const finalPlaceholders = finalProviders.filter(p => String(p.id).includes('placeholder'));
-		assert.strictEqual(finalPlaceholders.length, 1, 'Should have one remaining placeholder');
-
-		const authenticatedProvider = finalProviders.find(p => !String(p.id).includes('placeholder'));
-		assert.ok(authenticatedProvider, 'Should have one authenticated provider');
-	});
-});
